move buffer size detection code to Util.cpp (GetStringBufSize) fix compiling with -DDEBUG -D_DEBUG && -DUNICODE && -D__WX_DEBUG__
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/branches/x86_64@3609 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ src/os/linux/xsendstring.cpp
@@ -343,7 +343,7 @@ void pws_os::SendString(const char* str, AutotypeMethod method, unsigned delayMS
 
 	if (*str) {
 		/* some of the input chars were unprocessed */
-		sprintf(atGlobals.errorString, "char at approximate index(%u), ascii(%d), symbol(%c) couldn't be converted to keycode\n", ndest, (int)*str, *str);
+		sprintf(atGlobals.errorString, "char at approximate index(%u), ascii(%d), symbol(%c) couldn't be converted to keycode\n", static_cast<unsigned int>(ndest), (int)*str, *str);
 	}
 	else {
 		size_t n;
